split filter fuzz class
