+++ test/extensions/filters/http/common/fuzz/BUILD
@@ -3,6 +3,7 @@ licenses(["notice"])  # Apache 2
 load(
     "//bazel:envoy_build_system.bzl",
     "envoy_cc_fuzz_test",
+    "envoy_cc_test_library",
     "envoy_package",
     "envoy_proto_library",
 )
@@ -22,6 +23,20 @@ envoy_proto_library(
     ],
 )
 
+envoy_cc_test_library(
+    name = "uber_filter_lib",
+    hdrs = ["uber_filter.h"],
+    deps = [
+        ":filter_fuzz_proto_cc_proto",
+        "//source/common/config:utility_lib",
+        "//source/common/protobuf:utility_lib",
+        "//test/fuzz:utility_lib",
+        "//test/mocks/buffer:buffer_mocks",
+        "//test/mocks/http:http_mocks",
+        "//test/mocks/server:server_mocks",
+    ],
+)
+
 envoy_cc_fuzz_test(
     name = "filter_fuzz_test",
     srcs = ["filter_fuzz_test.cc"],
@@ -29,16 +44,9 @@ envoy_cc_fuzz_test(
     # All Envoy extensions must be linked to the test in order for the fuzzer to pick
     # these up via the NamedHttpFilterConfigFactory.
     deps = [
-        ":filter_fuzz_proto_cc_proto",
+        ":uber_filter_lib",
         "//source/common/config:utility_lib",
-        "//source/common/config:version_converter_lib",
         "//source/common/protobuf:utility_lib",
         "//test/config:utility_lib",
-        "//test/fuzz:utility_lib",
-        "//test/mocks/buffer:buffer_mocks",
-        "//test/mocks/http:http_mocks",
-        "//test/mocks/server:server_mocks",
-        "@envoy_api//envoy/service/auth/v3:pkg_cc_proto",
-        "@envoy_api//envoy/service/auth/v2alpha:pkg_cc_proto",
     ] + envoy_all_extensions(),
 )
